IdMap unit'y test with assignments (and add `but-testsupport::Sandbox`)
This also makes the Sandbox more generally available, in support of a more
generalised testing style of somewhat higher-level functions that use the context.
