+++ crates/but/src/command/legacy/rub/mod.rs
@@ -155,9 +155,9 @@ fn makes_no_sense_error(source: &CliId, target: &CliId) -> String {
     format!(
         "Operation doesn't make sense. Source {} is {} and target {} is {}.",
         source.to_string().blue().underline(),
-        source.kind().yellow(),
+        source.kind_for_humans().yellow(),
         target.to_string().blue().underline(),
-        target.kind().yellow()
+        target.kind_for_humans().yellow()
     )
 }
 
@@ -183,7 +183,7 @@ fn ids(
                         format!("{} (commit {})", id, &oid.to_string()[..7])
                     }
                     CliId::Branch { name, .. } => format!("{id} (branch '{name}')"),
-                    _ => format!("{} ({})", id, id.kind()),
+                    _ => format!("{} ({})", id, id.kind_for_humans()),
                 })
                 .collect();
             return Err(anyhow::anyhow!(
@@ -226,7 +226,7 @@ pub(crate) fn parse_sources(
                             format!("{} (commit {})", id, &oid.to_string()[..7])
                         }
                         CliId::Branch { name, .. } => format!("{id} (branch '{name}')"),
-                        _ => format!("{} ({})", id, id.kind()),
+                        _ => format!("{} ({})", id, id.kind_for_humans()),
                     })
                     .collect();
                 return Err(anyhow::anyhow!(
