+++ crates/but/tests/but/command/rub.rs
@@ -1,13 +1,13 @@
 use snapbox::str;
 
-use crate::utils::{Sandbox, setup_metadata};
+use crate::utils::Sandbox;
 
 #[test]
 fn shorthand_without_subcommand() -> anyhow::Result<()> {
     let env = Sandbox::init_scenario_with_target_and_default_settings("two-stacks")?;
 
     // Must set metadata to match the scenario
-    setup_metadata(&env, &["A", "B"])?;
+    env.setup_metadata(&["A", "B"])?;
 
     // Test that calling `but <id1> <id2>` defaults to rub
     // This should fail with a CliId not found error rather than a command not found error
