+++ crates/gitbutler-project/tests/add_cwd_as_project.rs
@@ -9,7 +9,7 @@ fn current_directory_dot() -> anyhow::Result<()> {
     // Change to the repository directory and add "." as the path
     std::env::set_current_dir(repo_path)?;
 
-    let project = gitbutler_project::add_with_path(tmp.path(), ".")?.unwrap_project();
+    let project = gitbutler_project::add_at_app_data_dir(tmp.path(), ".")?.unwrap_project();
 
     let expected_title = repo_path.file_name().unwrap().to_str().unwrap();
     assert_eq!(
